ignoring MSVC warnings: - warning C4625: copy constructor was implicitly defined as deleted - warning C4626: assignment operator was implicitly defined as deleted - warning C5027: move assignment operator was implicitly defined as deleted - warning C5026: move constructor was implicitly defined as deleted - warning C4514: unreferenced inline function has been removed - warning C4711: function selected for automatic inline expansion - warning C4710: function not inlined - warning C5045: Compiler will insert Spectre mitigation for memory load if /Qspectre switch specified - warning C4946: reinterpret_cast used between related classes: 'tesseract::ELIST2_LINK' and 'tesseract::WERD' - warning C5038: data member X will be initialized after data member Y
- warning C4623: default constructor was implicitly defined as deleted

fixing/marking MSVC warnings:
- as per the comments in the code.
- partially:
  + warning C5267: definition of implicit assignment operator is deprecated because it has a user-provided destructor
  + warning C5267: definition of implicit copy constructor is deprecated because it has a user-provided destructor
  + warning C5246: 'anonymous struct or union': the initialization of a subobject should be wrapped in braces
- warning C5262: implicit fall-through occurs here; are you missing a break statement? Use [[fallthrough]] when a break statement is intentionally omitted between cases
- warning C4800: Implicit conversion from 'int' to bool. Possible information loss  (Applies to code using several BoolParam instances, btw)
- warning C4355: 'this': used in base member initializer list
- warning C4855: implicit capture of 'this' via '[=]' is deprecated in '/std:c++20'
- warning C5263: calling 'std::move' on a temporary object prevents copy elision
- warning C4296: '<': expression is always false

# Conflicts:
#	src/api/altorenderer.cpp
#	src/api/hocrrenderer.cpp
#	src/api/pagerenderer.cpp
#	src/ccstruct/blob_bounds_calculator.cpp
#	src/ccutil/memcost_estimate.cpp
#	src/ccutil/thread_pool.hpp
#	src/classify/intmatcher.h
#	src/dict/permdawg.cpp
#	src/dict/trie.cpp
#	src/lstm/lstmrecognizer.cpp
#	src/lstm/lstmrecognizer.h
#	src/training/unicharset/normstrngs.cpp
#	src/viewer/scrollview.h
#	src/wordrec/lm_pain_points.cpp

+++ src/ccutil/scanutils.cpp
@@ -164,7 +164,8 @@ static double streamtofloat(FILE *s) {
       k *= 10;
     }
   }
-  double f = v + static_cast<double>(w) / k;
+  // warning C5219 : implicit conversion from 'uint64_t' to 'double', possible loss of data
+  double f = double(v) + double(w) / double(k);
   if (c == 'e' || c == 'E') {
     c = fgetc(s);
     int expsign = 1;
